edit for other server

+++ SwifflingBot.py
@@ -98,26 +98,27 @@ starttime = time.time()
 async def on_message(message):
     server = client.get_guild(413113734303580171)
     bot = discord.utils.get(server.members, name="Switchlings Bot")
-    q = message.content
-    for aword in allowedwords:
-        if aword in q.lower():
-            indexa = q.lower().index(aword)
-            indexb = indexa+len(aword)
-            zyx = q.split(q[indexa:indexb+1])
-            p = ""
-            for x in zyx:
-                p = "{}{}".format(p,x)
-            q = p
-    for word in badwords1:
-        if word in q.lower() and message.author != bot:#and noexception:
-            await message.channel.send("{}, Please don't joke about sensitive topics. It could lead to a perm ban. If you're serious about this, don't hesitate to DM a Switchling and they can help you.".format(message.author.mention))
-            await message.delete()
-            return
-    for word in badwords2:
-        if word in (q.lower()) and message.author != bot:# and noexception:
-            await message.channel.send("{}, Please don't joke about sensitive topics. It could lead to a perm ban. If you're serious about this, don't hesitate to DM a Switchling and they can help you.".format(message.author.mention))
-            await message.delete()
-            return                
+    if message.guild == server:
+        q = message.content
+        for aword in allowedwords:
+            if aword in q.lower():
+                indexa = q.lower().index(aword)
+                indexb = indexa+len(aword)
+                zyx = q.split(q[indexa:indexb+1])
+                p = ""
+                for x in zyx:
+                    p = "{}{}".format(p,x)
+                q = p
+        for word in badwords1:
+            if word in q.lower() and message.author != bot:#and noexception:
+                await message.channel.send("{}, Please don't joke about sensitive topics. It could lead to a perm ban. If you're serious about this, don't hesitate to DM a Switchling and they can help you.".format(message.author.mention))
+                await message.delete()
+                return
+        for word in badwords2:
+            if word in (q.lower()) and message.author != bot:# and noexception:
+                await message.channel.send("{}, Please don't joke about sensitive topics. It could lead to a perm ban. If you're serious about this, don't hesitate to DM a Switchling and they can help you.".format(message.author.mention))
+                await message.delete()
+                return                
     await client.process_commands(message)
     return
